+++ src/lib/netlist/plib/palloc.h
@@ -204,7 +204,7 @@ private:
 		char *data;
 	};
 
-	int new_block();
+	size_t new_block();
 
 	struct info
 	{
@@ -213,14 +213,14 @@ private:
 	};
 
 public:
-	mempool(int min_alloc, int min_align);
+	mempool(size_t min_alloc, size_t min_align);
 	~mempool();
 
 	void *alloc(size_t size);
 	void free(void *ptr);
 
-	int m_min_alloc;
-	int m_min_align;
+	size_t m_min_alloc;
+	size_t m_min_align;
 
 	std::vector<block> m_blocks;
 };
